Refactor INC quantization passes  (#249)
## Describe your changes
Move INC quantization passes from `olive.passes.onnx.quantization` to
`olive.passes.onnx.inc_quantization` to make them easier to maintain.
Also add external data config to the INC passes.

Update INC examples with the latest `engine.target` config option. 

## Checklist before requesting a review
- [ ] Add unit tests for this change.
- [ ] Make sure all tests can pass.
- [ ] Update documents if necessary.
- [ ] Format your code by running `pre-commit run --all-files`

## (Optional) Issue link
