+++ examples/bert_inc_ptq_cpu/bert_inc_static_config.json
@@ -44,8 +44,7 @@
                         "batch_size": 1
                     }
                 }
-            ],
-            "target": "local_system"
+            ]
         }
     },
     "passes": {
@@ -76,6 +75,7 @@
         },
         "evaluator": "common_evaluator",
         "host": "local_system",
+        "target": "local_system",
         "clean_cache": true,
         "cache_dir": "cache"
     }
