+++ olive/passes/onnx/__init__.py
@@ -5,18 +5,12 @@
 from olive.passes.onnx.append_pre_post_processing_ops import AppendPrePostProcessingOps
 from olive.passes.onnx.conversion import OnnxConversion
 from olive.passes.onnx.float16_conversion import OnnxFloatToFloat16
+from olive.passes.onnx.inc_quantization import IncDynamicQuantization, IncQuantization, IncStaticQuantization
 from olive.passes.onnx.insert_beam_search import InsertBeamSearch
 from olive.passes.onnx.mixed_precision import OrtMixedPrecision
 from olive.passes.onnx.model_optimizer import OnnxModelOptimizer
 from olive.passes.onnx.perf_tuning import OrtPerfTuning
-from olive.passes.onnx.quantization import (
-    IncDynamicQuantization,
-    IncQuantization,
-    IncStaticQuantization,
-    OnnxDynamicQuantization,
-    OnnxQuantization,
-    OnnxStaticQuantization,
-)
+from olive.passes.onnx.quantization import OnnxDynamicQuantization, OnnxQuantization, OnnxStaticQuantization
 from olive.passes.onnx.transformer_optimization import OrtTransformersOptimization
 
 __all__ = [
